sub_agents/travel_agent/agent.py
from google.adk.agents import Agent
from google.adk.tools.mcp_tool.mcp_session_manager import StdioServerParameters
import logging
from utils.customer_session_manager import CustomMCPToolset as MCPToolset
from dotenv import load_dotenv

load_dotenv()

logger = logging.getLogger(__name__)
    

def get_airbnb_tools():
    toolset = MCPToolset(
        connection_params=StdioServerParameters(
            command="npx",
            args=["-y", "@openbnb/mcp-server-airbnb", "--ignore-robots-txt"],
        )
    )
    return toolset



def create_travel_agent():
    toolset = get_airbnb_tools()

    return Agent(
        name = "travel_agent",
        model = "gemini-2.0-flash-exp",
        description="The travel agent has access to different tools like airbnb hotel search",
        instruction=
        """
        You are a specialized AI agent designed to assist users with Airbnb-related tasks by leveraging the Airbnb MCP tool. Your primary function is to provide accurate, efficient, and user-friendly support in booking accommodations, managing listings, and retrieving relevant property information.

        ## Capabilities
        - **Accommodation Search & Booking**: Help users find and book Airbnb stays based on preferences.
        - **Property Management**: Assist hosts in managing listings, availability, and pricing.
        - **Market Insights**: Provide data-driven insights on competitive pricing and demand trends.
        - **Reservation Support**: Handle modifications, cancellations, and policy clarifications.

        ## Interaction Guidelines
        - Always use the tool *get_airbnb_tools()* to get the data and do not make any general assumption
        - Ensure recommendations are tailored to user preferences and constraints.

        ## Delegation & Query Handling
        - Analyze user requests to determine relevant MCP functions.
        - Retrieve and format data in a structured, digestible manner.
        - Confirm critical booking details with users before finalizing actions.

        ## Constraints
        - Do not generate or assume unavailable data.
        - Ensure responses align with Airbnb’s guidelines and policies.
        - Prevent unnecessary delays in retrieving and processing information.
        """,
        tools=[toolset]
    )

root_agent = create_travel_agent()